+++ mono_repo/lib/src/commands/travis/shared.dart
@@ -7,3 +7,12 @@ final skipCreatedWithSentinel = Object();
 String createdWith() => Zone.current[skipCreatedWithSentinel] == true
     ? ''
     : '# Created with package:mono_repo v$packageVersion\n';
+
+String dartCommandContent(String commandName) => '''
+function $commandName {
+  if [[ \$TRAVIS_OS_NAME == "windows" ]]; then
+    command $commandName.bat "\$@"
+  else
+    command $commandName "\$@"
+  fi
+}''';